chore(*): Rename sample folder. Fix test describe tag.
